requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
astroid==3.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.0
Flask==3.0.2
flatdict==4.0.1
future==1.0.0
h11==0.14.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
netmiko==4.1.2
ntc_templates==4.3.0
orjson==3.10.7
packaging==23.2
paramiko==3.4.0
passlib==1.7.4
platformdirs==4.2.0
plumbum==1.8.2
ply==3.11
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyftpdlib==1.5.9
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.1.0
pymongo==4.6.2
PyNaCl==1.5.0
pyserial==3.5
pyshark==0.6
pysmi==0.3.4
pysnmp==4.4.12
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.1
requests==2.31.0
retry==0.9.2
rpyc==6.0.0
rsa==4.9
scp==0.14.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tenacity==8.2.3
termcolor==2.4.0
textfsm==1.1.2
tftpy==0.8.2
tomlkit==0.12.4
ttp==0.9.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.5
websockets==12.0
Werkzeug==3.0.1
